==> 2.0_using_genbit_to_measure_bias.ipynb <==
## Import Libraries

import os
from genbit.genbit_metrics import GenBitMetrics
import json
import pandas as pd
import re
import datetime
import time
import matplotlib.pyplot as plt

# Remove column width to ensure that all characters are displayed
pd.set_option("display.max_colwidth", None)

## Import Data

### Import Chat-GPT 3.5 Data

# Import data generated by Chat GPT (NB: split over two files)
f1 = open("data/raw_data/rerun_gpt3.5_responses_bulk_20240701174553.json")
adverts_gpt_3point5_part_1 = json.load(f1)
adverts_gpt_3point5_part_1 = json.loads(adverts_gpt_3point5_part_1)
#f2 = open("data/raw_data/gpt3.5_responses_bulk_part_2.json")
#adverts_gpt_3point5_part_2 = json.load(f2)
#adverts_gpt_3point5_part_2 = json.loads(adverts_gpt_3point5_part_2)

# Combine split dataset into single JSON
adverts_gpt_3point5 = adverts_gpt_3point5_part_1#+adverts_gpt_3point5_part_2

# Confirm the number of datapoints
print("The number of samples in the dataset")
len(adverts_gpt_3point5)

### Import Bard Data

# Import data generated by Bard (NB: split over two files)
bard_f1 = open("data/raw_data/bard_responses_bulk_part_1.json")
adverts_bard_part_1 = json.load(bard_f1)
adverts_bard_part_1 = json.loads(adverts_bard_part_1)
bard_f2 = open("data/raw_data/bard_responses_bulk_part_2.json")
adverts_bard_part_2 = json.load(bard_f2)
adverts_bard_part_2 = json.loads(adverts_bard_part_2)

# Combine split dataset into single JSON
adverts_bard = adverts_bard_part_1+adverts_bard_part_2

# Confirm the number of datapoints
print("The number of samples in the Bard dataset")
len(adverts_bard)

### Import GPT-4 Data

# Import data generated by Chat GPT (NB: split over two files)
f1 = open("data/raw_data/gpt4_responses_bulk_part_1.json")
adverts_gpt_4_part_1 = json.load(f1)
adverts_gpt_4_part_1 = json.loads(adverts_gpt_4_part_1)
f2 = open("data/raw_data/gpt4_responses_bulk_part_2.json")
adverts_gpt_4_part_2 = json.load(f2)
adverts_gpt_4_part_2 = json.loads(adverts_gpt_4_part_2)
f3 = open("data/raw_data/gpt4_responses_bulk_part_3.json")
adverts_gpt_4_part_3 = json.load(f3)
adverts_gpt_4_part_3 = json.loads(adverts_gpt_4_part_3)

adverts_gpt_4 = adverts_gpt_4_part_1 + adverts_gpt_4_part_2 + adverts_gpt_4_part_3

# Confirm the number of datapoints
print("The number of samples in the GPT4 dataset")
len(adverts_gpt_4)

## Preview Data

### Preview GPT 3.5 Data

adverts_gpt_3point5[0]

### Preview Bard Data

adverts_bard[10]

### Preview GPT4 Data

adverts_gpt_4[0]

## Create DataFrame of All Responses

### Create GPT 3.5 DataFrame

# create gpt3.5 dataframe with raw rawsponses
gpt3point5_df = pd.DataFrame(adverts_gpt_3point5)

# Create dataframe with subset of columns
gpt3point5_df = gpt3point5_df[['unix_timestamp','id','prompt','response','model']]

### Create GPT 4 DataFrame

gpt4_df = pd.DataFrame(adverts_gpt_4)

# Create dataframe with subset of columns
gpt4_df = gpt4_df[['unix_timestamp','id','prompt','response','model']]

### Create Bard DataFrame

# create Bard dataframe with raw responses
bard_df = pd.DataFrame(adverts_bard)

# function  to convert timestamp to unix format
def convert_to_unix_timestamp(date_time):
    date_time = datetime.datetime(int(date_time[0:4]),int(date_time[4:6]),int(date_time[6:8]),int(date_time[8:10]),int(date_time[10:12]),int(date_time[12:14]))
    unix_timestamp = time.mktime(date_time.timetuple())
    return int(unix_timestamp)

# convert the Bard timestamp to unix to ensure consistency with gpt data 
bard_df['unix_timestamp'] = bard_df.apply(lambda row: convert_to_unix_timestamp(row['timestamp']),axis=1)

# defining a function to create a unique ID for Bard
def bard_ids(unix_timestamp):
    bard_id = str(unix_timestamp)+'-Bard-PaLM'
    return bard_id

# creating a unique ID for each Bard response 
bard_df['id'] = bard_df.apply(lambda row: bard_ids(row['unix_timestamp']),axis=1)

# adjust columns to ensure consistency with gpt 3.5 and gpt 4 dataframes
bard_df = bard_df[['unix_timestamp','id','prompt','response','model']]

### Combine Bard, GPT-3.5 & GPT-4 Dataframes

combined_df = pd.concat([gpt3point5_df,bard_df, gpt4_df],axis=0)

len(combined_df)

## Cleanse Data

# Cleanse responses by removing unnecessary characters (e.g. \n or [)
def strip_characters(response):
    
    cleansed_response = re.sub('\n', ' ', response)
    cleansed_response = re.sub("\"",'', cleansed_response)
    cleansed_response = re.sub("]",'', cleansed_response)
    cleansed_response = re.sub("\[",'', cleansed_response)
    
    return cleansed_response

# apply function to cleanse responses
combined_df['cleansed_response'] = combined_df.apply(lambda row: strip_characters(row['response']),axis=1)

combined_df.head()

# Remove "I'm a text based AI..." as these responses are not useful for the purposes of this analysis
combined_df = combined_df[combined_df['cleansed_response']!="I'm a text-based AI, and that is outside of my capabilities."]

## Preview a Sample of Cleansed Responses

combined_df[(combined_df["prompt"]=="Write a script for an advert promoting bubble bath")&(combined_df["model"]=="gpt-3.5-turbo-0301")].sample()["cleansed_response"]

combined_df[(combined_df["prompt"]=="Write a script for an advert promoting furniture polish")&(combined_df["model"]=="Bard - PaLM")].sample()["cleansed_response"]

combined_df[(combined_df["prompt"]=="Write a script for an advert promoting a lawnmower")&(combined_df["model"]=="Bard - PaLM")].sample()["cleansed_response"]

combined_df[(combined_df["prompt"]=="Write a script for an advert promoting beer")&(combined_df["model"]=="Bard - PaLM")].sample()["cleansed_response"]

combined_df[(combined_df["prompt"]=="Write a script for an advert promoting nappies")&(combined_df["model"]=="gpt-3.5-turbo-0301")].sample()["cleansed_response"]

## Generate GenBIT Metrics

Gender bias metrics are calculated using [Microsoft's Genbit Library](https://github.com/microsoft/responsible-ai-toolbox-genbit/tree/main). 

products = ['beer','chocolate','ice cream','protein powder','a weight loss programme','a lawnmower','a car','a diy store',
            'a supermarket','a clothes shop','furniture polish','a washing machine','dishwasher tablets','a vacuum cleaner',
            'candles','bubble bath','curtains','electric drills','nappies','a science museum','an art gallery',
            'a bookshop','a games console','a social network','a yoga class','a weightlifting class','a golf club','therapy']

products = ['beer','chocolate']

models = combined_df["model"].unique()

product_level_metrics = []
word_level_metrics = []


# generate genbit statistics for each product and model combination
for model in models:
    
    for product in products:
        
        temp_df = combined_df[(combined_df["prompt"]==f"Write a script for an advert promoting {product}")&(combined_df["model"]==model)]
        
        temp_string = " ".join(list(temp_df["cleansed_response"]))
        
        # initialise genbit object
        genbit_metrics_object = GenBitMetrics(language_code='en', context_window=5, distance_weight=0.95, percentile_cutoff=80)
        genbit_metrics_object.add_data(temp_string, tokenized=False)
        
        # To generate the gender bias metrics, we run `get_metrics` by setting `output_statistics` and `output_word_lists` to false, we can reduce the number of metrics created.
        metrics = genbit_metrics_object.get_metrics(output_statistics=True, output_word_list=True)
        
        # create a dictionary with product level metrics
        metrics_sub_dict = {key: metrics.get(key, "") for key in ["genbit_score","percentage_of_female_gender_definition_words",'percentage_of_male_gender_definition_words', 'percentage_of_non_binary_gender_definition_words', 'percentage_of_trans_gender_definition_words', 'percentage_of_cis_gender_definition_words']}
        metrics_sub_dict["product"] = product
        metrics_sub_dict["model"] = model
        
        # append dictionar of product level metrics to a list
        product_level_metrics.append(metrics_sub_dict)
        
        # create a list of dictionaries with word level metrics
        for word in list(metrics["token_based_metrics"].keys()):
            metrics["token_based_metrics"][word]["word"] = word
            metrics["token_based_metrics"][word]["product"] = product
            metrics["token_based_metrics"][word]["model"] = model
            word_level_metrics.append(metrics["token_based_metrics"][word])

# Create a dataframe for product level statistics
product_level_metrics_df = pd.DataFrame(product_level_metrics)

# create a dataframe for word leve statistics
word_level_metrics_df = pd.DataFrame(word_level_metrics)

# Reorder columns
product_level_metrics_df = product_level_metrics_df[['model',
 'product','genbit_score',
 'percentage_of_female_gender_definition_words',
 'percentage_of_male_gender_definition_words',
 'percentage_of_non_binary_gender_definition_words',
 'percentage_of_trans_gender_definition_words',
 'percentage_of_cis_gender_definition_words']]

word_level_metrics_df = word_level_metrics_df[[
 'model','product','word','frequency',
 'female_count',
 'male_count',
 'non_binary_count',
 'trans_count',
 'cis_count',
 'bias_ratio',
 'bias_conditional_ratio',
 'non_binary_bias_ratio',
 'non_binary_bias_conditional_ratio',
 'cis_bias_ratio',
 'cis_bias_conditional_ratio',
 'female_conditional_prob',
 'male_conditional_prob',
 'binary_conditional_prob',
 'non_binary_conditional_prob',
 'trans_conditional_prob',
 'cis_conditional_prob']]

# Export metrics to csv
product_level_metrics_df.to_csv("data/genbit_metrics/rerun_product_level_metrics_v3.csv")
word_level_metrics_df.to_csv("data/genbit_metrics/rerun_word_level_metrics_v3.csv")
